+++ CHANGELOG.md
@@ -95,6 +95,7 @@ in the naming of release branches.
 - The `Cardano.Ledger.Alonzo.Scripts` module no longer re-exports the
   `plutus-ledger-api`'s `assertWellFormedCostModelParams`. #3065
 - Removed unused `Data.BiMap` module from `cardano-data` #3089
+- Removed `getMultiSigBytes` as unused #?
 
 ### Fixed